adding Rubriques

+++ src/main/java/fr/ubo/dosi/projectagile/cscievaebackend/DTO/RubriqueDTO.java
@@ -20,7 +20,7 @@ public class RubriqueDTO {
     @NotNull
     @Size(max = 32)
     private String designation;
-    private Short ordre;
+    private Long ordre;
     private Set<QuestionDTO> questionEvaluations = new LinkedHashSet<>();
